started tag filter, added active-inactive style

+++ app/Http/Controllers/TagController.php
@@ -7,6 +7,11 @@ use Illuminate\Http\Request;
 
 class TagController extends Controller
 {
+
+    public function __construct()
+    {
+        $this->middleware('auth');
+    }
     /**
      * Display a listing of the resource.
      *
@@ -89,8 +94,9 @@ class TagController extends Controller
      */
     public function destroy(Tag $tag)
     {
-        $tag->delete();
-
+        if (count($tag->notes) == 0) {
+            $tag->delete();
+        }
         return redirect()->back();
     }
 }
